refactor(formatter): split markup.rs into markup/ module
Pure move: markup.rs (2937 lines) becomes markup/mod.rs plus ten
submodules. Function bodies and comments move verbatim; only `use`
statements and item visibility change, and lib.rs is untouched because
mod.rs re-exports the crate-visible symbols at their original paths.

Verified with a coverage script: the declared line ranges tile the
original 1..2937 exactly once with no gap or overlap, and each range
appears verbatim in its target file. The digest over 4193 corpus files
is unchanged: {"digest": "1720de4d97fda965", "ok": 4193, "error": 0,
"panic": 0}.
